Change all masks to type torch.bool (#3890)
* get_text_field_mask

* masked_softmax

* masked_log_softmax

* masked_max/mean

* get_lengths_from_binary_sequence_mask

* get_final_encoder_states

* replace_masked_values

* batched_span_select

* sequence_cross_entropy_with_logits

* add_sentence_boundary_token_ids/remove_sentence_boundaries

* scalar mix

* More changes

* Fix tests

* More changes, mostly in tests

* Test fix

* black

* More changes

* More cahnges

+++ allennlp/models/basic_classifier.py
@@ -115,7 +115,7 @@ class BasicClassifier(Model):
             A scalar loss to be optimised.
         """
         embedded_text = self._text_field_embedder(tokens)
-        mask = get_text_field_mask(tokens).float()
+        mask = get_text_field_mask(tokens)
 
         if self._seq2seq_encoder:
             embedded_text = self._seq2seq_encoder(embedded_text, mask=mask)
